fix: RootLayout, Main moved to page, navigation folder for related components
